Add initHighlightId to PdfHighlight props
This is used in the last version of react-pdf-highlighter to scroll
directly to an highlight if the book is loaded with the hash in the link
at first.

If an initHighlightId is provided then we load the pdf and scroll to the
said highlight. We do not scroll to the init page number which would be
the behavior if we did not provide a highlight to scroll to in the first
place.